+++ Core/EntityExtensions.cs
@@ -155,7 +155,7 @@ namespace Scellecs.Morpeh {
 #endif
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static void AddTransfer(this Entity entity, int typeId) {
+        internal static void AddTransfer(this Entity entity, long typeId) {
             if (entity.previousArchetypeId == -1) {
                 entity.previousArchetypeId = entity.currentArchetypeId;
             }
@@ -170,7 +170,7 @@ namespace Scellecs.Morpeh {
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static void RemoveTransfer(this Entity entity, int typeId) {
+        internal static void RemoveTransfer(this Entity entity, long typeId) {
             if (entity.previousArchetypeId == -1) {
                 entity.previousArchetypeId = entity.currentArchetypeId;
             }
